# src/main/java/com/bancobci/entrevista/controller/UsuarioController.java
package com.bancobci.entrevista.controller;


import com.bancobci.entrevista.dto.UsuarioDto;
import com.bancobci.entrevista.service.AuthenticationService;
import org.slf4j.Logger;
import org.slf4j.LoggerFactory;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.DeleteMapping;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PatchMapping;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.PutMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

import java.util.List;

@RestController
@RequestMapping("/api/usuarios")
public class UsuarioController {
    private static final Logger log = LoggerFactory.getLogger(UsuarioController.class);
    private final AuthenticationService authenticationService;
    public UsuarioController(AuthenticationService authenticationService) {
        this.authenticationService = authenticationService;
    }
    @GetMapping
    public ResponseEntity<List<UsuarioDto>> getAllUsuarios() {
        List<UsuarioDto> allUsuarios = authenticationService.getAllUsuarios();
        return new ResponseEntity<>(allUsuarios, HttpStatus.OK);
    }
    @GetMapping("/{id}")
    public ResponseEntity<?> getUsuarioById(@PathVariable Long id) {
        UsuarioDto usuarioById = authenticationService.getUsuarioById(id);
        if (usuarioById != null)
            return new ResponseEntity<>(usuarioById, HttpStatus.OK);
        return new ResponseEntity<>(HttpStatus.NOT_FOUND);
    }
    @PutMapping("/{id}")
    public ResponseEntity<UsuarioDto> update(@PathVariable Long id,
                                             @RequestBody UsuarioDto usuarioDTO){
        log.info("email: {}", usuarioDTO.getEmail());
        UsuarioDto usuarioActualizado = authenticationService.updateUsuario(id, usuarioDTO);
        if (usuarioActualizado != null) {
            return new ResponseEntity<>(usuarioActualizado, HttpStatus.OK);
        } else {
            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
        }
    }
    @PatchMapping("/{id}")
    public ResponseEntity<UsuarioDto> patch(@PathVariable Long id,
                                            @RequestBody UsuarioDto usuarioDTO){
        log.info("PATCH usuario id: {}", id);
        UsuarioDto usuarioActualizado = authenticationService.patchUsuario(id, usuarioDTO);
        if (usuarioActualizado != null) {
            return new ResponseEntity<>(usuarioActualizado, HttpStatus.OK);
        } else {
            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
        }
    }
    @DeleteMapping("/{id}")
    public ResponseEntity<Void> delete(@PathVariable Long id){
        authenticationService.deleteUsuario(id);
        return new ResponseEntity<>(HttpStatus.NO_CONTENT);
    }
}