# src/main/java/com/bancobci/entrevista/dto/UsuarioDto.java
package com.bancobci.entrevista.dto;

import jakarta.validation.constraints.Email;
import jakarta.validation.constraints.NotBlank;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

import java.util.List;


@Getter
@Setter
@Builder
@AllArgsConstructor
@NoArgsConstructor
public class UsuarioDto {
    @NotBlank
    private String nombre;
    @Email(regexp = "^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\\.[a-zA-Z]{2,}$", message = "El correo electrónico es inválido")
    private String email;
    private String password;
    private List<PhoneDto> phones;
    private List<TareaDto> tareas;
}